streamlit project/requirements.txt
streamlit
numpy
pandas
plotly.graph_objects
plotly.express
matplotlib.pyplot
warnings